rewrite representation to a sorted binary list and embed it
This is now a binary format which is embeded as a string, this remove the need to run code at init or when lazy initializing.
It also fixes a bug where we would incorrectly assume the dataset only reports CIDR alligned ranges, however it does not.

This result in a `1.56MiB` binary file for the dataset.
It is embedded as a string into the `rodata` part of the executable. This remove the runtime memory cost (as this is usually mmaped directly from disk) and all the init execution. It also removes some dependencies, in practice here is the difference (tested with `go test -c`):
```console
> ls -l old new
-rwxr-xr-x 1 hugo hugo 10398323 Dec 19 02:40 old*
-rwxr-xr-x 1 hugo hugo  6954814 Dec 19 02:39 new*
```

This is also significantly faster, new:
```
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-libp2p-asn-util
cpu: AMD Ryzen 5 3600 6-Core Processor
BenchmarkAsnForIPv6-12    	21599720	        56.85 ns/op
```
old:
```
BenchmarkAsnForIPv6-12    	 9417902	       126.3 ns/op
```